Summary of Desired Outcome
Product SKUs (comma-separated):
99166948
 Product Source URL:
https://www.grundfos.com/za

This code allows the user to input a list of product SKUs and a URL where the information for the products can be found. When the user submits the form, the JavaScript code sends a POST request to the server with the SKUs and source URL, which can then be processed by the PHP code to fetch the product information and create the products. Note: The implementation of the server-side code to fetch the product information and create the products based on the SKUs and source URL is beyond the scope of this answer and would require additional development.